extract CommentIndicator
