+++ client/app/reader/PdfListView.jsx
@@ -6,7 +6,6 @@ import { bindActionCreators } from 'redux';
 import Table from '../components/Table';
 import { formatDateStr } from '../util/DateUtil';
 import Comment from '../components/Comment';
-import Button from '../components/Button';
 import { linkToSingleDocumentView } from '../components/PdfUI';
 import DocumentCategoryIcons from '../components/DocumentCategoryIcons';
 import DocumentListHeader from '../components/reader/DocumentListHeader';
@@ -86,52 +85,6 @@ const lastReadIndicatorMapStateToProps = (state, ownProps) => ({
   shouldShow: state.ui.pdfList.lastReadDocId === ownProps.docId
 });
 const ConnectedLastReadIndicator = connect(lastReadIndicatorMapStateToProps)(LastReadIndicator);
-
-class CommentIndicator extends React.PureComponent {
-  toggleComments = () => {
-    this.props.handleToggleCommentOpened(this.props.doc.id);
-  }
-
-  render() {
-    const numberOfComments = _.size(this.props.numberAnnotations);
-    const icon = `fa fa-3 ${this.props.doc.listComments ?
-      'fa-angle-up' : 'fa-angle-down'}`;
-    const name = `expand ${numberOfComments} comments`;
-
-    return <span className="document-list-comments-indicator">
-      {numberOfComments > 0 &&
-        <span>
-          <Button
-            classNames={['cf-btn-link']}
-            href="#"
-            ariaLabel={name}
-            name={name}
-            id={`expand-${this.props.doc.id}-comments-button`}
-            onClick={this.toggleComments}>{numberOfComments}
-            <i className={`document-list-comments-indicator-icon ${icon}`}/>
-          </Button>
-        </span>
-      }
-    </span>;
-  }
-}
-
-const commentIndicatorMapStateToProps = (state, ownProps) => ({
-  numberAnnotations: getAnnotationByDocumentId(state, ownProps.doc.id)
-});
-const commentIndicatorMapDispatchToProps = (dispatch) => ({
-  handleToggleCommentOpened(docId) {
-    dispatch({
-      type: Constants.TOGGLE_COMMENT_LIST,
-      payload: {
-        docId
-      }
-    });
-  }
-});
-const ConnectedCommentIndicator = connect(
-  commentIndicatorMapStateToProps,
-  commentIndicatorMapDispatchToProps)(CommentIndicator);
 
 export class PdfListView extends React.Component {
   constructor() {
